% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomes}
\alias{biomes}
\title{Ontology of biomes}
\format{
A data.frame (337x3) containing nested biomes and habitats
}
\usage{
biomes
}
\description{
A hierarchical list of biomes and synonyms as grouped by the IUCN Habitats
Classification Scheme Version 3.1.
}
\keyword{datasets}
