% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conservation_actions}
\alias{conservation_actions}
\title{Ontology of insect conservation actions}
\format{
A data.frame (523x7) containing nested conservation actions
}
\usage{
conservation_actions
}
\description{
A hierarchical list of insect conservation actions ranging from general
approaches (e.g. "Land/water management") to specific actions (e.g. "use
glazing treatments to reduce light spill from inside lit buildings).
}
\keyword{datasets}
