% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dictionary.R
\name{create_dictionary}
\alias{create_dictionary}
\title{Creates hierarchical dictionary entries}
\usage{
create_dictionary(x, return_dictionary = TRUE)
}
\arguments{
\item{x}{a data.frame where columns represent parent and child entries in a hierarchy from left to right}

\item{return_dictionary}{logical; if TRUE, returns a dictionary object; if FALSE, returns a list of lists}
}
\value{
a list of lists or hierarchical dictionary object of parents, children, and descendants
}
\description{
Given an object containing parent and child entries in columns from left to right, creates hierarchical named lists of lists that can be used as a dictionary object
}
