% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_terms.R
\name{define_terms}
\alias{define_terms}
\title{Define terms within a dictionary}
\usage{
define_terms(entry, exclusion_terms = NULL)
}
\arguments{
\item{entry}{a character vector of entries in a dictionary}

\item{exclusion_terms}{a character vector of terms to not retain as possible parts of a term definition}
}
\value{
a named list of the same length as entry containing character vectors defining each entry
}
\description{
Extracts phrases associated with a dictionary entry to define it while removing exclusion terms that are too common or irrelevant to be included in definitions
}
