% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_smartly.R
\name{error_rates}
\alias{error_rates}
\title{Gets model classification error rates}
\usage{
error_rates(input, output)
}
\arguments{
\item{input}{a character vector of known tags for testing model}

\item{output}{a character vector of tags from model predictions}
}
\value{
a numeric score of model accuracy
}
\description{
Given known data and model predictions, calculates model classification accuracy
}
