% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_levels.R
\name{extract_levels}
\alias{extract_levels}
\title{Extract tags at specified levels of an ontology}
\usage{
extract_levels(x, n.levels = NULL)
}
\arguments{
\item{x}{a character vector containing hierarchical tags separated by a period}

\item{n.levels}{numeric: the indices of levels that should be extracted}
}
\value{
a list of length equal to the length of n.levels where the length of each entry is the length of x
}
\description{
Given the results of tagging documents with a hierarchical ontology, exports the tags from user-specified levels of the original ontology.
}
