% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_rows.R
\name{fill_rows}
\alias{fill_rows}
\title{Fill in parents for nested entries}
\usage{
fill_rows(x)
}
\arguments{
\item{x}{a data.frame where columns represent, from left to right, nested levels of a hierarchical list}
}
\value{
a data.frame of the same dimensions as x with parent entries filled in
}
\description{
Given a human-readable nested list, matches children and descendants to parent entries and fills in gaps.
}
\examples{

}
