% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_smartly.R
\name{get_tags}
\alias{get_tags}
\title{Get tags from model predictions}
\usage{
get_tags(predictions, cutoff, modfam, tags)
}
\arguments{
\item{predictions}{a vector or matrix of model predictions}

\item{cutoff}{numeric: a cutoff in probability for accepting predictions as tags}

\item{modfam}{string: either "binomial" or "multinomial"}

\item{tags}{factor: tags predicted by the model}
}
\description{
Get tags from model predictions
}
