% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dictionary.R
\name{make_entry}
\alias{make_entry}
\title{Creates hierarchical dictionary entries}
\usage{
make_entry(parents, children, descendants = NULL)
}
\arguments{
\item{parents}{a character vector of parent entries}

\item{children}{a character vector of length equal to parents containing child entries}

\item{descendants}{a named list of descendants of child entries (optional)}
}
\value{
a list of lists or hierarchical dictionary object of parents, children, and descendants
}
\description{
Given a list of parent and children entries, creates hierarchical named lists of lists that can be used as a dictionary object
}
