% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine_terms.R
\name{mine_terms}
\alias{mine_terms}
\title{Extracts reasonable terms for topic modeling}
\usage{
mine_terms(x, known_phrases = NULL, ngrams = TRUE, new_stopwords = c())
}
\arguments{
\item{x}{a character vector of documents}

\item{known_phrases}{a character vector of stopwords to retain or phrases known to be relevant}

\item{ngrams}{logical: if TRUE, only returns phrases of two words or more}

\item{new_stopwords}{a character vector of additional, subject-specific terms to remove}
}
\description{
Given documents, extracts possible ngram terms to use for topic modeling.
}
