% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pest_articles}
\alias{pest_articles}
\title{Example pest studies dataset}
\format{
A data.frame with 300 rows of 11 variables:
\describe{
\item{title}{the article title}
\item{abstract}{the article abstract}
\item{keywords}{database-tagged keywords}
\item{authors}{article authors}
\item{source}{journal title or other source of the article}
\item{year}{year of publication}
\item{volume}{volume of publication}
\item{issue}{issue of publication}
\item{startpage}{start page for article in publication}
\item{endpage}{end page for article in publication}
\item{doi}{the article Digital Object Identifier (DOI)}
}
}
\usage{
pest_articles
}
\description{
A dataset containing a set of example articles on crop pests
}
\keyword{datasets}
