% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_smartly.R
\name{prep_matrix}
\alias{prep_matrix}
\title{Prep sparse matrix for use in tag_smartly}
\usage{
prep_matrix(words)
}
\arguments{
\item{words}{a list of character vectors of words}
}
\value{
a sparse document-feature matrix
}
\description{
Given a list of bags of words, creates a document-feature matrix
}
