% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_freely.R
\name{tag_freely}
\alias{tag_freely}
\title{Group documents by unknown topic}
\usage{
tag_freely(docs, k = 3, ngrams = TRUE, n_terms = 10)
}
\arguments{
\item{docs}{a character vector of documents}

\item{k}{numeric: how many groups should be created}

\item{ngrams}{logical: if TRUE, groups documents using ngrams (i.e. multi-word phrases) instead of single terms}

\item{n_terms}{numeric: how many terms should be returned for each topic}
}
\value{
a vector of topic numbers and associated terms for users to post-hoc classify
}
\description{
Assign topics to documents when there is no scheme or prior knowledge for classifying documents other than number of topics.
}
\examples{

}
