% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_smartly.R
\name{tag_smartly}
\alias{tag_smartly}
\title{Tag topics based on known documents}
\usage{
tag_smartly(
  new_documents,
  tagged_documents,
  tags,
  cutoff = NULL,
  prop_training = 0.7
)
}
\arguments{
\item{new_documents}{a character vector of documents to tag}

\item{tagged_documents}{a character vector of documents}

\item{tags}{a character vector of tags of the same length as tagged_documents}

\item{cutoff}{numeric: what cutoff should be used for probability of being included in a topic? If NULL, finds the optimal cutoff for model accuracy in training data.}

\item{prop_training}{numeric: what proportion of tagged_documents should be used for training? Remaining proportion will be used for model testing.}
}
\value{
a character vector of tags of length equal to new_documents
}
\description{
Given a set of documents tagged with topics and a set of new documents, tags new documents by learning characteristics of topics from existing tagged documents.
}
