% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_strict.R
\name{tag_strict}
\alias{tag_strict}
\title{Applies an ontology to tag document metadata}
\usage{
tag_strict(doc, scheme, enable_stemming = TRUE, allow_multiple = TRUE)
}
\arguments{
\item{doc}{character: a vector of document texts to tag}

\item{scheme}{a dictionary object or list containing an ontology or set of terms}

\item{enable_stemming}{logical: if TRUE, interpret lemmatized stems of words as synonymous (e.g. "burning" and "burned" are equivalent to "burn")}

\item{allow_multiple}{logical: if TRUE, returns all matched metadata, else returns most frequent}
}
\description{
Given a hierarchical ontology or
dictionary and a set of documents, applies the
ontoloy to tag metadata hierarchically for each document.
}
