% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine_terms.R
\name{mine_terms}
\alias{mine_terms}
\title{Extracts reasonable terms for topic modeling}
\usage{
mine_terms(x, known_phrases = NULL, ngrams = TRUE, new_stopwords = c())
}
\arguments{
\item{x}{a character vector of documents}

\item{known_phrases}{a character vector of stopwords to retain or phrases known to be relevant}

\item{ngrams}{logical: if TRUE, only returns phrases of two words or more}

\item{retain_stopwords}{logical: if TRUE, retains stopwords used in known terms}
}
\description{
Given documents, extracts possible ngram terms to use for topic modeling.
}
\examples{
doc <- "Recently burned coniferous forests host wildlife communities 
        that respond to variation in burn severity, post-fire habitat 
        structure, and patch configuration. Habitat selection theory 
        predicts that birds inhabiting these variable post-fire landscapes 
        will select nesting locations that confer an adaptive advantage 
        through increased fitness and reproductive success. Understanding 
        the effect of post-fire habitat on avian nesting ecology can 
        provide valuable information to guide restoration and management 
        after wildfire. The Black-backed Woodpecker (Picoides arcticus) 
        is strongly associated with recently burned forests in the western 
        United States, where it is used as an indicator species for the 
        effects of post-fire forest management."

keywords <- c("protection of habitat")

mine_terms(doc, retain_stopwords=TRUE, known_phrases=keywords)
}
