% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine_terms.R
\name{mine_terms}
\alias{mine_terms}
\title{Extracts reasonable terms for topic modeling}
\usage{
mine_terms(x, retain_stopwords = FALSE, known_phrases = NULL, ngrams = TRUE)
}
\arguments{
\item{x}{a character vector of documents}

\item{retain_stopwords}{logical: if TRUE, retains stopwords used in known terms}

\item{known_phrases}{a character vector of stopwords to retain or phrases known to be relevant}
}
\description{
Given documents, extracts possible ngram terms to use for topic modeling.
}
\examples{
inst/examples/mine_terms.R
}
