% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_strictly.R
\name{should_stem}
\alias{should_stem}
\title{Checks length of a word and stems if long enough to be unambiguous}
\usage{
should_stem(word)
}
\description{
Checks length of a word and stems if long enough to be unambiguous
}
