% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_strictly.R
\name{tag_strictly}
\alias{tag_strictly}
\title{Applies an ontology to tag document metadata}
\usage{
tag_strictly(doc, scheme, allow_multiple = TRUE)
}
\arguments{
\item{doc}{character: a vector of document texts to tag}

\item{scheme}{a dictionary object or list containing an ontology or set of terms}

\item{allow_multiple}{logical: if TRUE, returns all matched metadata, else returns most frequent}
}
\description{
Given a hierarchical ontology or
dictionary and a set of documents, applies the
ontoloy to tag metadata hierarchically for each document.
}
